Moved test setup into conftest.
